requirements.txt
requests==2.21.0
slackclient==2.0.0
websocket==0.2.1
websocket-client==0.56.0
Flask==1.0.2
python-dotenv==0.10.1
gunicorn==19.9.0